first part of work with server
