+++ src/actions/auth.js
@@ -1,4 +1,6 @@
 import * as types from '../constants';
+import * as routes from '../constants/routes';
+import callApi from '../utils/call-api';
 
 export function signup(username, password) {
 	return (dispatch) => {
@@ -6,24 +8,12 @@ export function signup(username, password) {
 			type: types.SIGNUP_REQUEST
 		});
 
-		return fetch('http://localhost:8000/v1/signup', {
-			method: 'POST',
-			headers: {
-				'Accept': 'application/json',
-				'Content-Type': 'application/json'
-			},
-			body: JSON.stringify({
-				username,
-				password
-			})
-		})
-		.then(response => response.json())
-		.then(json => {
-			if (json.success) {
-				return json;
-			}
-			throw new Error(json.message);
-		})
+		return callApi(
+			routes.SIGNUP,
+			null,
+			{ method: 'POST' },
+			{ username, password }
+		)
 		.then(json => {
 			if (!json.token) {
 				throw new Error('Token has not been provided!');
@@ -49,24 +39,12 @@ export function login(username, password) {
 			type: types.LOGIN_REQUEST
 		});
 
-		return fetch('http://localhost:8000/v1/login', {
-			method: 'POST',
-			headers: {
-				'Accept': 'application/json',
-				'Content-Type': 'application/json'
-			},
-			body: JSON.stringify({
-				username,
-				password
-			})
-		})
-		.then(response => response.json())
-		.then(json => {
-			if (json.success) {
-				return json;
-			}
-			throw new Error(json.message);
-		})
+		return callApi(
+			routes.LOGIN,
+			null,
+			{ method: 'POST' },
+			{ username, password }
+		 )
 		.then(json => {
 			if (!json.token) {
 				throw new Error('Token has not been provided!');
@@ -105,20 +83,12 @@ export function recieveAuth() {
 			})
 		}
 
-		return fetch('http://localhost:8000/v1/users/me', {
-			headers: {
-				'Authorization': `Bearer ${token}`,
-				'Accept': 'application/json',
-				'Content-Type': 'application/json'
-			},
-		})
-		.then(response => response.json())
-		.then(json => {
-			if (json.success) {
-				return json;
-			}
-			throw new Error(json.message);
-		})
+		return callApi(
+			routes.RECIEVE_AUTH,
+			token,
+			{},
+			{}
+		)
 		.then(json => dispatch({
 				type: types.RECIEVE_AUTH_SUCCESS,
 				payload: json
