+++ src/constants/routes.js
@@ -1,0 +1,5 @@
+export const ROOT_ROUTE = Symbol('http://localhost:8000/v1/');
+
+export const LOGIN = Symbol('login');
+export const SIGNUP = Symbol('signup');
+export const RECIEVE_AUTH = Symbol('users/me');
