+++ src/constants/services.js
@@ -1,0 +1,1 @@
+export const REDIRECT = Symbol('REDIRECT');
